+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 * Update the **GeT-RM WGS** tutorial to include more genes and also GRCh38.
 * Make the profiles (e.g. copy number) look prettier.
 * Add CNV callers for GRCh38.
+* Rename :meth:`sdk.utils.Archive.check` method to :meth:`sdk.utils.Archive.check_type`.
 
 0.8.0 (2021-11-20)
 ------------------